feat(re-encryption): start example
The example aims to encrpypt a message, and then make it available for each subscriber to decrypt the message using their own private key. The proxy re-encryption is required here to wrap the symmetric key the message was encrypted with and distribute it to subscribers in capsulas that can only be opened with their private keys.